Boat images + win/loss + turns
boats now have better visual representation, and you can land a hit for a win or a loss against your opponent! Boats are spaced about 3 screen widths apart so that shooting is more tactical/trial and error. Shooting is now turn based, and the host begins. The other player gets a turn once the cannon ball fired by the last has come to a stop.